server-side caching implemented

+++ src/services/postgres/NotesService.js
@@ -6,9 +6,10 @@ const AuthorizationError = require('../../exceptions/AuthorizationError');
 const { mapDBToModel } = require('../../utils');
 
 class NotesService {
-  constructor(collaborationsService) {
+  constructor(collaborationsService, cacheService) {
     this.pool = new Pool();
     this.collaborationsService = collaborationsService;
+    this.cacheService = cacheService;
   }
 
   /**
@@ -34,19 +35,33 @@ class NotesService {
       throw new InvariantError('Catatan gagal ditambahkan');
     }
 
+    await this.cacheService.delete(`notes:${owner}`)
+
     return result.rows[0].id;
   }
 
   async getNotes(owner) {
-    const query = {
-      text: `SELECT notes.* FROM notes
-      LEFT JOIN collaborations ON collaborations.note_id = notes.id
-      WHERE notes.owner = $1 OR collaborations.user_id = $1
-      GROUP BY notes.id`,
-      values: [owner],
-    };
-    const result = await this.pool.query(query);
-    return result.rows.map(mapDBToModel);
+    try {
+      // get the notes data from cache first
+      const result = await this.cacheService.get(`notes:${owner}`)
+      return JSON.parse(result)
+    } catch (error) {
+      // if failed, then get the notes data from database
+      const query = {
+        text: `SELECT notes.* FROM notes
+        LEFT JOIN collaborations ON collaborations.note_id = notes.id
+        WHERE notes.owner = $1 OR collaborations.user_id = $1
+        GROUP BY notes.id`,
+        values: [owner],
+      };
+      const result = await this.pool.query(query);
+      const mappedResult = result.rows.map(mapDBToModel);
+
+      // cache the result before returning the result
+      await this.cacheService.set(`notes:${owner}`, JSON.stringify(mappedResult))
+
+      return mappedResult;
+    }
   }
 
   async getNoteById(id) {
@@ -70,7 +85,7 @@ class NotesService {
   async editNoteById(id, { title, body, tags }) {
     const updatedAt = new Date().toISOString();
     const query = {
-      text: 'UPDATE notes SET title = $1, body = $2, tags = $3, updated_at = $4 WHERE id = $5 RETURNING id',
+      text: 'UPDATE notes SET title = $1, body = $2, tags = $3, updated_at = $4 WHERE id = $5 RETURNING id, owner',
       values: [title, body, tags, updatedAt, id],
     };
 
@@ -79,11 +94,15 @@ class NotesService {
     if (!result.rowCount) {
       throw new NotFoundError('Gagal memperbarui catatan. Id tidak ditemukan');
     }
+
+    const { owner } = result.rows[0]
+
+    await this.cacheService.delete(`notes:${owner}`)
   }
 
   async deleteNoteById(id) {
     const query = {
-      text: 'DELETE FROM notes WHERE id = $1',
+      text: 'DELETE FROM notes WHERE id = $1 RETURNING id, owner',
       values: [id],
     };
 
@@ -92,6 +111,10 @@ class NotesService {
     if (!result.rowCount) {
       throw new NotFoundError('Gagal menghapus catatan. Id tidak ditemukan');
     }
+
+    const { owner } = result.rows[0]
+
+    await this.cacheService.delete(`notes:${owner}`)
   }
 
   async verifyNoteOwner(id, owner) {
